requirements.txt
aiohttp==3.7.4.post0
aioschedule==0.5.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
black==19.10b0
boto3==1.17.97
botocore==1.20.97
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dateparser==1.0.0
discord.py==1.7.3
distlib==0.3.2
docutils==0.17.1
filelock==3.0.12
flake8==3.9.2
humanize==3.9.0
identify==2.2.10
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
mccabe==0.6.1
more-itertools==8.8.0
multidict==5.1.0
multitasking==0.0.9
nodeenv==1.6.0
numpy==1.20.3
openai==0.27.2
packaging==20.9
pandas==1.2.4
pandas-datareader==0.9.0
pathspec==0.8.1
Pillow==8.2.0
pluggy==0.13.1
pre-commit==2.13.0
pur==5.4.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
riotwatcher==3.1.1
s3transfer==0.4.2
scipy==1.6.3
six==1.16.0
toml==0.10.2
tqdm==4.65.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
virtualenv==20.4.7
wcwidth==0.2.5
websockets==9.1
yarl==1.6.3
yfinance==0.1.59
